+++ ChangeLog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Add new subcommands `app` and `device` under import and export subcommands
 - The documentation about the different available `buffer_id` has been updated to provide the complete list for each compatible device.
 - add `O3R_ODS_RENDERED_ZONES` `buffer_id`
 - Add `ovp8xx`,`o3x1xx` and `o3d3xx` and `discover` commands as a primary level subcommands to ifm3d command line interface